+++ CHANGELOG.md
@@ -16,6 +16,13 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 
 ### Fixes and improvements
 
+## [1.17.1](https://github.com/OpenNMT/OpenNMT-tf/releases/tag/v1.17.1) (2019-01-21)
+
+### Fixes and improvements
+
+* Fix offset in alignment vectors introduced by the `<s>` special token
+* Fix crash when using the AdafactorOptimizer and setting beta1
+
 ## [1.17.0](https://github.com/OpenNMT/OpenNMT-tf/releases/tag/v1.17.0) (2019-01-10)
 
 ### New features